Add cmake and vcpkg named module providers
This patch implements a named module provider using cmake.
Because a named provider is eager and cmake doens't provider a list of all
installed packages, the provider requires a pre-defined list of modules to
check.
The actual patch provides support for the Boost and the Poco modules.

The patch also implements a named provider for the vcpkg. It uses the cmake
module provider by just setting the toolchain file provided by vcpkg.

Fixes: QBS-1041
Change-Id: I492e5a345575f304035804960751994d3a5be8c1